+++ maps-kt-geojson/src/commonMain/kotlin/center/sciprog/maps/geojson/GeoJsonPropertiesAttribute.kt
@@ -1,7 +1,0 @@
-package center.sciprog.maps.geojson
-
-import kotlinx.serialization.json.JsonObject
-import kotlinx.serialization.serializer
-import space.kscience.SerializableAttribute
-
-public object GeoJsonPropertiesAttribute : SerializableAttribute<JsonObject>("properties", serializer())